move chassis-specific tests to chassis_test
